Move visual and audio cue into their own windows
